+++ src/danmaku_sender/core/validator_session.py
@@ -2,7 +2,7 @@ import copy
 import logging
 from typing import Any, TypedDict
 
-from .services.validator import ValidationIssue, validate_danmaku_list
+from .services.danmaku_validator import ValidationIssue, validate_danmaku_list
 from .models.danmaku import Danmaku
 from .state import AppState
